feat: major refactor to TypeScript with new color system
- Migrate entire codebase from JavaScript to TypeScript
- Implement new Yellow+Blue/Purple+Green color system
- Add comprehensive TypeScript types and interfaces
- Refactor components to use new design system
- Update documentation with new color specifications
- Add new project overview documentation
- Improve component architecture and type safety
- Update build configuration for TypeScript
- Add proper error boundaries and theme providers
- Enhance photography gallery with better type safety
- Fix linting errors for unused variables

This represents a significant upgrade to the codebase architecture and design system.
