+++ PROJECT_OVERVIEW.md
@@ -1,0 +1,52 @@
+# Project Overview: The Abhiyan Sainju Portfolio Website
+
+**Version:** 1.0  
+**Date:** September 25, 2025  
+**Owner:** Abhiyan Sainju
+
+## 1. Project Goal & Vision
+
+The goal is to design, develop, and deploy a unique, professional portfolio website that serves as a personal brand hub. The site will showcase Abhiyan Sainju's identity as a skilled **Software Engineer** (projects, experience), a creative **Photographer** (gallery with EXIF details), and a critical **Thinker** (deep dives, journey from Kathmandu → DC).
+
+This project moves beyond a simple digital resume to combine professionalism, creativity, and personality into a memorable site that recruiters and professional connections will remember.
+
+## 2. Technology & Tools
+
+| Category         | Tool / Technology | Purpose                                                                 |
+| ---------------- | ----------------- | ----------------------------------------------------------------------- |
+| Framework        | React (with Vite) | For building a modern, fast, component-based user interface.            |
+| Styling          | Tailwind CSS      | For a utility-first, fully custom, and responsive design system.      |
+| Animations       | Framer Motion     | To add polished, professional, and fluid animations.                    |
+| Deployment       | Vercel            | For seamless, continuous deployment connected to the Git repository.    |
+| Domain           | abhiyansainju.com | For a professional and branded online presence.                         |
+| Version Control  | Git & GitHub      | For disciplined code management and collaboration.                      |
+| Project Tracking | GitHub Projects   | For task and progress tracking using a Kanban board.                    |
+
+## 3. Phased Deliverables & Timeline
+
+The project will be executed in four distinct phases for iterative development.
+
+*   **Phase 1: Foundation & Brand Identity (MVP)** (Week 1)
+*   **Phase 2: Core Content & Interactivity** (Weeks 2-3)
+*   **Phase 3: The "Uniquely Abhiyan" Layer** (Week 4)
+*   **Phase 4: Final Polish & Deployment** (Week 5)
+
+## 4. Git Workflow & Versioning
+
+*   **Branching Strategy:** A simplified GitFlow model (`main`, `develop`, `feat/*`, `fix/*`).
+*   **Pull Requests (PRs):** All features will be merged into `develop` via a PR.
+*   **Commit Style:** Conventional Commits.
+*   **Releases:** Semantic Versioning.
+
+## 5. Project Documentation
+
+*   [**Requirements Document**](./REQUIREMENTS.md)
+*   [**Design Document**](./DESIGN_DOCUMENT.md)
+*   [**Design System**](./DESIGN_SYSTEM.md)
+
+## 6. Future Enhancements (Post-Launch)
+
+*   **Abhiyan Bot:** A playful, interactive chatbot to answer FAQs.
+*   **"Now" Section:** A page to share what you are currently learning or working on.
+*   **Photography × Code Fusion:** An experimental section to overlay data visualizations on photo backgrounds.
+*   **Football Easter Egg:** A hidden animation for a touch of personality.
