+++ index.html
@@ -2,7 +2,7 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" type="image/x-icon" href="/favicon.ico" />
+    <link rel="icon" type="image/svg+xml" href="/favicon.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Abhiyan Sainju – Software Engineer & Photographer</title>
     <meta name="description" content="Software Engineer (Cloud + AI) and Photographer. Projects, deep dives, and photography by Abhiyan Sainju." />
