--- src/App.jsx
@@ -1,30 +1,0 @@
-import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
-import Layout from "./components/Layout";
-import Home from "./pages/Home";
-import Projects from "./pages/Projects";
-import Photography from "./pages/Photography";
-import About from "./pages/About";
-import DeepDives from "./pages/DeepDives";
-import NotFound from "./pages/NotFound";
-import { ThemeProvider } from "./contexts/ThemeContext";
-
-function App() {
-  return (
-    <ThemeProvider>
-      <Router>
-        <Layout>
-          <Routes>
-            <Route path="/" element={<Home />} />
-            <Route path="/projects" element={<Projects />} />
-            <Route path="/photography" element={<Photography />} />
-            <Route path="/about" element={<About />} />
-            <Route path="/deep-dives" element={<DeepDives />} />
-            <Route path="*" element={<NotFound />} />
-          </Routes>
-        </Layout>
-      </Router>
-    </ThemeProvider>
-  );
-}
-
-export default App;
