+++ src/contexts/useTheme.ts
@@ -1,0 +1,12 @@
+import { useContext } from 'react';
+import { ThemeContext } from './ThemeContext';
+import { ThemeContextType } from './Theme';
+
+// Custom hook to use theme context
+export const useTheme = (): ThemeContextType => {
+  const context = useContext(ThemeContext);
+  if (context === undefined) {
+    throw new Error('useTheme must be used within a ThemeProvider');
+  }
+  return context;
+};
