+++ src/data/articles.ts
@@ -1,0 +1,80 @@
+export const articles = [
+    {
+      id: 1,
+      title:
+        "AI Copyright Law: Navigating the Legal Landscape of Machine Learning",
+      abstract:
+        "An in-depth analysis of copyright implications in AI training, fair use doctrine, and the evolving legal framework surrounding machine learning models.",
+      date: "March 2024",
+      readTime: "12 min read",
+      category: "AI & Law",
+      featured: true,
+      slug: "ai-copyright-law",
+    },
+    {
+      id: 2,
+      title: "The Coupon Collector's Problem: A Mathematical Journey",
+      abstract:
+        "Exploring one of probability theory's most elegant problems, from basic combinatorics to advanced statistical applications in real-world scenarios.",
+      date: "February 2024",
+      readTime: "8 min read",
+      category: "Mathematics",
+      featured: false,
+      slug: "coupon-collectors-problem",
+    },
+    {
+      id: 3,
+      title: "Cloud Architecture Patterns: Building Scalable Systems",
+      abstract:
+        "A comprehensive guide to designing resilient, scalable cloud architectures using modern patterns and best practices from production environments.",
+      date: "January 2024",
+      readTime: "15 min read",
+      category: "Cloud Computing",
+      featured: true,
+      slug: "cloud-architecture-patterns",
+    },
+    {
+      id: 4,
+      title: "The Psychology of Code Reviews: Fostering Better Collaboration",
+      abstract:
+        "Understanding the human side of technical reviews, communication patterns, and strategies for creating more effective development teams.",
+      date: "December 2023",
+      readTime: "10 min read",
+      category: "Software Engineering",
+      featured: false,
+      slug: "psychology-code-reviews",
+    },
+    {
+      id: 5,
+      title:
+        "Photography as Data: Computational Analysis of Visual Composition",
+      abstract:
+        "Exploring how machine learning can analyze photographic composition, color theory, and aesthetic principles to enhance visual storytelling.",
+      date: "November 2023",
+      readTime: "14 min read",
+      category: "AI & Photography",
+      featured: false,
+      slug: "photography-computational-analysis",
+    },
+    {
+      id: 6,
+      title: "The Future of DevOps: AI-Driven Infrastructure Management",
+      abstract:
+        "Examining how artificial intelligence is transforming infrastructure automation, predictive maintenance, and deployment strategies.",
+      date: "October 2023",
+      readTime: "11 min read",
+      category: "DevOps",
+      featured: true,
+      slug: "ai-driven-devops",
+    },
+  ];
+
+export const categories = [
+    "All",
+    "AI & Law",
+    "Mathematics",
+    "Cloud Computing",
+    "Software Engineering",
+    "AI & Photography",
+    "DevOps",
+  ];
