+++ tsconfig.node.json
@@ -1,0 +1,16 @@
+{
+  "compilerOptions": {
+    "composite": true,
+    "skipLibCheck": true,
+    "module": "ESNext",
+    "moduleResolution": "bundler",
+    "allowSyntheticDefaultImports": true
+  },
+  "include": [
+    "vite.config.js",
+    "commitlint.config.js",
+    "eslint.config.js",
+    "tailwind.config.js",
+    "scripts/**/*.js"
+  ]
+}
